Fixed a bug with renewals, and added sleep support

+++ cleanup.sh
@@ -1,13 +1,15 @@
 #!/bin/bash
 
-if [ ! -s key.txt ]; then
-        echo "Key file is empty or does not exist";
-        exit 1;
+SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
+
+if [ ! -s `echo ${SCRIPT_DIR}/key.txt` ]; then
+	echo "Key file is empty or does not exist"
+	exit 1
 fi
 
-if [ ! -s zone.txt ]; then
-        echo "Zone file is empty or does not exist";
-        exit 1;
+if [ ! -s `echo ${SCRIPT_DIR}/zone.txt` ]; then
+	echo "Zone file is empty or does not exist"
+	exit 1
 fi
 
 if ! command -v jq &> /dev/null; then
@@ -20,8 +22,13 @@ if ! command -v curl &> /dev/null; then
     exit 1
 fi
 
-TW_ZONE=`cat zone.txt`
-TW_KEY=`cat key.txt`
+if [ -z $CERTBOT_AUTH_OUTPUT ]; then
+	echo "No auth output. Are you running this script through certbot? (An example command is included in command.txt)"
+	exit 1
+fi
+
+TW_ZONE=`cat ${SCRIPT_DIR}/zone.txt`
+TW_KEY=`cat ${SCRIPT_DIR}/key.txt`
 TW_CERTBOT_RECORD_ID=`echo $CERTBOT_AUTH_OUTPUT | jq .domain_record.id`
 
 curl -s -X DELETE -H "Content-Type: application/json" -H "Authorization: Bearer $TW_KEY" "https://api.digitalocean.com/v2/domains/$TW_ZONE/records/$TW_CERTBOT_RECORD_ID"
